+++ validation/validator_test.go
@@ -1,9 +1,7 @@
 package validation
 
 import (
-	"net/http/httptest"
 	"reflect"
-	"strings"
 	"testing"
 
 	"github.com/System-Glitch/goyave/v2/lang"
@@ -107,107 +105,89 @@ func (suite *ValidatorTestSuite) TestAddRule() {
 }
 
 func (suite *ValidatorTestSuite) TestValidate() {
-	rawRequest := httptest.NewRequest("POST", "/test-route", strings.NewReader("string=hello%20world&number=42"))
-	rawRequest.Header.Set("Content-Type", "application/x-www-form-urlencoded; param=value")
-	errors := Validate(rawRequest, nil, RuleSet{}, "en-US")
+	errors := Validate(nil, RuleSet{}, false, "en-US")
 	suite.Equal(1, len(errors))
 	suite.Equal("Malformed request", errors["error"][0])
 
-	rawRequest = httptest.NewRequest("POST", "/test-route", strings.NewReader("string=hello%20world&number=42"))
-	rawRequest.Header.Set("Content-Type", "application/json")
-	errors = Validate(rawRequest, nil, RuleSet{}, "en-US")
+	errors = Validate(nil, RuleSet{}, true, "en-US")
 	suite.Equal(1, len(errors))
 	suite.Equal("Malformed JSON", errors["error"][0])
 
-	rawRequest = httptest.NewRequest("POST", "/test-route", strings.NewReader("string=hello%20world&number=42"))
-	rawRequest.Header.Set("Content-Type", "application/json")
-	errors = Validate(rawRequest, map[string]interface{}{
+	errors = Validate(map[string]interface{}{
 		"string": "hello world",
 		"number": 42,
 	}, RuleSet{
 		"string": {"required", "string"},
 		"number": {"required", "numeric", "min:10"},
-	}, "en-US")
+	}, true, "en-US")
 	suite.Equal(0, len(errors))
 
-	rawRequest = httptest.NewRequest("POST", "/test-route", strings.NewReader("string=hello%20world&number=42"))
-	rawRequest.Header.Set("Content-Type", "application/json")
 	data := map[string]interface{}{
 		"nullField": nil,
 	}
-	errors = Validate(rawRequest, data, RuleSet{
+	errors = Validate(data, RuleSet{
 		"nullField": {"numeric"},
-	}, "en-US")
+	}, true, "en-US")
 	_, exists := data["nullField"]
 	suite.False(exists)
 	suite.Equal(0, len(errors))
 
-	rawRequest = httptest.NewRequest("POST", "/test-route", strings.NewReader("string=hello%20world&number=42"))
-	rawRequest.Header.Set("Content-Type", "application/json")
 	data = map[string]interface{}{
 		"nullField": nil,
 	}
-	errors = Validate(rawRequest, data, RuleSet{
+	errors = Validate(data, RuleSet{
 		"nullField": {"required", "nullable", "numeric"},
-	}, "en-US")
+	}, true, "en-US")
 	_, exists = data["nullField"]
 	suite.True(exists)
 	suite.Equal(0, len(errors))
 
-	rawRequest = httptest.NewRequest("POST", "/test-route", strings.NewReader("string=hello%20world&number=42"))
-	rawRequest.Header.Set("Content-Type", "application/json")
 	data = map[string]interface{}{
 		"nullField": "test",
 	}
-	errors = Validate(rawRequest, data, RuleSet{
+	errors = Validate(data, RuleSet{
 		"nullField": {"required", "nullable", "numeric"},
-	}, "en-US")
+	}, true, "en-US")
 	_, exists = data["nullField"]
 	suite.True(exists)
 	suite.Equal(1, len(errors))
 }
 
 func (suite *ValidatorTestSuite) TestValidateWithArray() {
-	rawRequest := httptest.NewRequest("POST", "/test-route", strings.NewReader("string=hello"))
-	rawRequest.Header.Set("Content-Type", "application/x-www-form-urlencoded")
 	data := map[string]interface{}{
 		"string": "hello",
 	}
-	errors := Validate(rawRequest, data, RuleSet{
+	errors := Validate(data, RuleSet{
 		"string": {"required", "array"},
-	}, "en-US")
+	}, false, "en-US")
 	suite.Equal("array", GetFieldType(data["string"]))
 	suite.Equal("hello", data["string"].([]string)[0])
 	suite.Equal(0, len(errors))
 }
 
 func (suite *ValidatorTestSuite) TestValidateArrayValues() {
-	rawRequest := httptest.NewRequest("POST", "/test-route", strings.NewReader("string=hello&string=world"))
-	rawRequest.Header.Set("Content-Type", "application/x-www-form-urlencoded")
 	data := map[string]interface{}{
 		"string": []string{"hello", "world"},
 	}
-	errors := Validate(rawRequest, data, RuleSet{
+	errors := Validate(data, RuleSet{
 		"string": {"required", "array", ">min:3"},
-	}, "en-US")
+	}, false, "en-US")
 	suite.Equal(0, len(errors))
 
-	rawRequest = httptest.NewRequest("POST", "/test-route", strings.NewReader("string=hi&string=world"))
-	rawRequest.Header.Set("Content-Type", "application/x-www-form-urlencoded")
 	data = map[string]interface{}{
 		"string": []string{"hi", ",", "there"},
 	}
-	errors = Validate(rawRequest, data, RuleSet{
+	errors = Validate(data, RuleSet{
 		"string": {"required", "array", ">min:3"},
-	}, "en-US")
+	}, false, "en-US")
 	suite.Equal(1, len(errors))
 
 	data = map[string]interface{}{
 		"string": []string{"johndoe@example.org", "foobar@example.org"},
 	}
-	errors = Validate(rawRequest, data, RuleSet{
+	errors = Validate(data, RuleSet{
 		"string": {"required", "array:string", ">email"},
-	}, "en-US")
+	}, true, "en-US")
 	suite.Equal(0, len(errors))
 
 	suite.Panics(func() {
@@ -215,26 +195,22 @@ func (suite *ValidatorTestSuite) TestValidateArrayValues() {
 	})
 
 	// Empty array
-	rawRequest = httptest.NewRequest("POST", "/test-route", strings.NewReader(""))
-	rawRequest.Header.Set("Content-Type", "application/x-www-form-urlencoded")
 	data = map[string]interface{}{
 		"string": []string{},
 	}
-	errors = Validate(rawRequest, data, RuleSet{
+	errors = Validate(data, RuleSet{
 		"string": {"array", ">uuid:5"},
-	}, "en-US")
+	}, true, "en-US")
 	suite.Equal(0, len(errors))
 }
 
 func (suite *ValidatorTestSuite) TestValidateTwoDimensionalArray() {
-	rawRequest := httptest.NewRequest("POST", "/test-route", strings.NewReader(""))
-	rawRequest.Header.Set("Content-Type", "application/x-www-form-urlencoded")
 	data := map[string]interface{}{
 		"values": [][]interface{}{{"0.5", 1.42}, {0.6, 7}},
 	}
-	errors := Validate(rawRequest, data, RuleSet{
+	errors := Validate(data, RuleSet{
 		"values": {"required", "array", ">array:numeric"},
-	}, "en-US")
+	}, false, "en-US")
 	suite.Equal(0, len(errors))
 
 	arr, ok := data["values"].([][]float64)
@@ -247,14 +223,12 @@ func (suite *ValidatorTestSuite) TestValidateTwoDimensionalArray() {
 		suite.Equal(7.0, arr[1][1])
 	}
 
-	rawRequest = httptest.NewRequest("POST", "/test-route", strings.NewReader(""))
-	rawRequest.Header.Set("Content-Type", "application/x-www-form-urlencoded")
 	data = map[string]interface{}{
 		"values": [][]float64{{5, 8}, {0.6, 7}},
 	}
-	errors = Validate(rawRequest, data, RuleSet{
+	errors = Validate(data, RuleSet{
 		"values": {"required", "array", ">array:numeric", ">min:3"},
-	}, "en-US")
+	}, true, "en-US")
 	suite.Equal(1, len(errors))
 
 	_, ok = data["values"].([][]float64)
@@ -263,40 +237,38 @@ func (suite *ValidatorTestSuite) TestValidateTwoDimensionalArray() {
 	data = map[string]interface{}{
 		"values": [][]float64{{5, 8, 6}, {0.6, 7, 9}},
 	}
-	errors = Validate(rawRequest, data, RuleSet{
+	errors = Validate(data, RuleSet{
 		"values": {"required", "array", ">array:numeric", ">min:3"},
-	}, "en-US")
+	}, true, "en-US")
 	suite.Equal(0, len(errors))
 
 	data = map[string]interface{}{
 		"values": [][]float64{{5, 8}, {3, 7}},
 	}
-	errors = Validate(rawRequest, data, RuleSet{
+	errors = Validate(data, RuleSet{
 		"values": {"required", "array", ">array:numeric", ">>min:3"},
-	}, "en-US")
+	}, true, "en-US")
 	suite.Equal(0, len(errors))
 
 	data = map[string]interface{}{
 		"values": [][]float64{{5, 8}, {0.6, 7}},
 	}
-	errors = Validate(rawRequest, data, RuleSet{
+	errors = Validate(data, RuleSet{
 		"values": {"required", "array", ">array:numeric", ">>min:3"},
-	}, "en-US")
+	}, true, "en-US")
 	suite.Equal(1, len(errors))
 }
 
 func (suite *ValidatorTestSuite) TestValidateNDimensionalArray() {
-	rawRequest := httptest.NewRequest("POST", "/test-route", strings.NewReader(""))
-	rawRequest.Header.Set("Content-Type", "application/x-www-form-urlencoded")
 	data := map[string]interface{}{
 		"values": [][][]interface{}{
 			{{"0.5", 1.42}, {0.6, 4, 3}},
 			{{"0.6", "1.43"}, {}, {2}},
 		},
 	}
-	errors := Validate(rawRequest, data, RuleSet{
+	errors := Validate(data, RuleSet{
 		"values": {"required", "array", ">array", ">>array:numeric", ">max:3", ">>>max:4"},
-	}, "en-US")
+	}, true, "en-US")
 	suite.Equal(0, len(errors))
 
 	arr, ok := data["values"].([][][]float64)
@@ -316,9 +288,9 @@ func (suite *ValidatorTestSuite) TestValidateNDimensionalArray() {
 			{{"0.6", "1.43"}, {}, {2}, {4}},
 		},
 	}
-	errors = Validate(rawRequest, data, RuleSet{
+	errors = Validate(data, RuleSet{
 		"values": {"required", "array", ">array", ">>array:numeric", ">max:3", ">>>max:4"},
-	}, "en-US")
+	}, true, "en-US")
 	suite.Equal(1, len(errors))
 
 	data = map[string]interface{}{
@@ -327,9 +299,9 @@ func (suite *ValidatorTestSuite) TestValidateNDimensionalArray() {
 			{{"0.6", "1.43"}, {}, {2}},
 		},
 	}
-	errors = Validate(rawRequest, data, RuleSet{
+	errors = Validate(data, RuleSet{
 		"values": {"required", "array", ">array", ">>array:numeric", ">max:3", ">>>max:4"},
-	}, "en-US")
+	}, true, "en-US")
 	suite.Equal(1, len(errors))
 }