+++ docs_src/src/guide/changelog.md
@@ -9,6 +9,8 @@
     - Can be [toggled at runtime](./advanced/multi-services#maintenance-mode).
     - The server can be started in maintenance mode using the `maintenance` config option. (Defaults to `false`)
 - Added [advanced array validation](./basics/validation#validating-arrays), with support for n-dimensional arrays.<Badge text="BETA" type="warn"/>
+- Malformed request messages can now be localized. (`malformed-request` and `malformed-json` entries in `locale.json`)
+- Modified the validator to allow [manual validation](./basics/validation#manual-validation).
 
 ## v2.0.0